Improve identify js docs
