--- frontend/common/code-help/install/install-rn.js
@@ -1,7 +1,0 @@
-module.exports = ({ NPM_CLIENT, NPM_RN_CLIENT }) => `// npm
-npm i ${NPM_CLIENT} ${NPM_RN_CLIENT} --save
-
-// yarn
-yarn add ${NPM_CLIENT} ${NPM_RN_CLIENT}
-
-`
