+++ frontend/common/code-help/traits/traits-react.js
@@ -1,9 +1,26 @@
 module.exports = (
   envId,
-  { FEATURE_NAME, FEATURE_NAME_ALT, LIB_NAME, TRAIT_NAME },
-  USER_ID,
+  { FEATURE_NAME, FEATURE_NAME_ALT, LIB_NAME, TRAIT_NAME, USER_ID },
+  userId,
 ) => `
-// Home Page
+// Option 1: Initialise with an identity and traits
+import { FlagsmithProvider } from 'flagsmith/react';
+
+export default function App() {
+  return (
+    &lt;FlagsmithProvider
+      options={{
+        environmentID: '${envId}',
+        identity: '${userId || USER_ID}',
+        traits: {${TRAIT_NAME}: 21},
+      }}
+      flagsmith={flagsmith}&gt;
+      {...Your app}
+    &lt;/FlagsmithProvider>
+  );
+}
+
+// Option 2: Set traits / identify after initialising
 import flagsmith from '${LIB_NAME}';
 import { useFlags, useFlagsmith } from 'flagsmith/react';
 
@@ -13,7 +30,9 @@ export default function HomePage() {
   const ${FEATURE_NAME_ALT} = flags.${FEATURE_NAME_ALT}.value
 
   const identify = () => {
-    flagsmith.identify('${USER_ID}', {${TRAIT_NAME}:21}); // only causes re-render if the user has overrides / segment overrides for ${FEATURE_NAME} or ${FEATURE_NAME_ALT}
+    flagsmith.identify('${
+      userId || USER_ID
+    }', {${TRAIT_NAME}:21}); // only causes re-render if the user has overrides / segment overrides for ${FEATURE_NAME} or ${FEATURE_NAME_ALT}
   };
 
   const setTrait = () => {
