+++ frontend/common/constants.ts
@@ -10,7 +10,6 @@ const keywords = {
   LIB_NAME_JAVA: 'FlagsmithClient',
   NPM_CLIENT: 'flagsmith',
   NPM_NODE_CLIENT: 'flagsmith-nodejs',
-  NPM_RN_CLIENT: 'react-native-flagsmith',
   SEGMENT_NAME: 'superUsers',
   TRAIT_NAME: 'age',
   URL_CLIENT: 'https://cdn.jsdelivr.net/npm/flagsmith/index.js',
@@ -18,7 +17,10 @@ const keywords = {
   USER_FEATURE_NAME: 'my_even_cooler_feature',
   USER_ID: 'user_123456',
 }
-
+const keywordsReactNative = {
+  ...keywords,
+  NPM_CLIENT: 'react-native-flagsmith',
+}
 export default {
   archivedTag: { color: '#8f8f8f', label: 'Archived' },
   codeHelp: {
@@ -73,9 +75,9 @@ export default {
         keywords,
         userId,
       ),
-      'React Native': require('./code-help/create-user/create-user-rn')(
+      'React Native': require('./code-help/create-user/create-user-react')(
         envId,
-        keywords,
+        keywordsReactNative,
         userId,
       ),
       'Ruby': require('./code-help/create-user/create-user-ruby')(
@@ -111,7 +113,10 @@ export default {
       'PHP': require('./code-help/init/init-php')(envId, keywords),
       'Python': require('./code-help/init/init-python')(envId, keywords),
       'React': require('./code-help/init/init-react')(envId, keywords),
-      'React Native': require('./code-help/init/init-rn')(envId, keywords),
+      'React Native': require('./code-help/init/init-js')(
+        envId,
+        keywordsReactNative,
+      ),
       'Ruby': require('./code-help/init/init-ruby')(envId, keywords),
       'Rust': require('./code-help/init/init-rust')(envId, keywords),
       'curl': require('./code-help/init/init-curl')(envId, keywords),
@@ -129,7 +134,9 @@ export default {
       'PHP': require('./code-help/install/install-php')(keywords),
       'Python': require('./code-help/install/install-python')(keywords),
       'React': require('./code-help/install/install-js')(keywords),
-      'React Native': require('./code-help/install/install-rn')(keywords),
+      'React Native': require('./code-help/install/install-js')(
+        keywordsReactNative,
+      ),
       'Ruby': require('./code-help/install/install-ruby')(keywords),
       'Rust': require('./code-help/install/install-rust')(keywords),
       'curl': require('./code-help/install/install-curl')(keywords),
@@ -199,9 +206,9 @@ export default {
         keywords,
         userId,
       ),
-      'React Native': require('./code-help/traits/traits-rn')(
+      'React Native': require('./code-help/traits/traits-react')(
         envId,
-        keywords,
+        keywordsReactNative,
         userId,
       ),
       'Ruby': require('./code-help/traits/traits-ruby')(
