+++ frontend/web/components/CodeHelp.js
@@ -1,6 +1,4 @@
 import React, { Component } from 'react'
-import Tabs from './base/forms/Tabs'
-import TabItem from './base/forms/TabItem'
 import Highlight from './Highlight'
 import ConfigProvider from 'common/providers/ConfigProvider'
 import Constants from 'common/constants'
